+++ ChangeLog.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 - improvements to markdown to ascii text renderer to allow the Venice
   Github flavoured markdown documentation to be converted to pure ascii
   to feed it as embeddings to LLMs
+- :cargo-qdrant module to install/start/stop Qdrant Vector DB docker 
+  containers 
   
 ### Changed